+++ examples/data-sources/meraki_network_policies_by_client/data-source.tf
@@ -1,0 +1,3 @@
+data "meraki_network_policies_by_client" "example" {
+  network_id = "L_123456"
+}