+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Add `meraki_organization_devices` data source
 - Add `meraki_organization_firmware_upgrades` data source
 - Add `meraki_organization_inventory_devices` data source
+- Add `meraki_network_policies_by_client` data source
 
 ## 0.1.2